fix(testing-sdk): retry operation missing NextAttemptTimestamp
