--- packages/aws-durable-execution-sdk-js-examples/src/examples/__tests__/promise-any.test.ts.skip
@@ -1,39 +1,0 @@
-import { LocalDurableTestRunner} from "@aws/durable-execution-sdk-js-testing";
-import { handler } from '../promise-any';
-
-// TODO: enable test when SDK lib supports retries in concurrent operations
-
-beforeAll(() => LocalDurableTestRunner.setupTestEnvironment());
-afterAll(() => LocalDurableTestRunner.teardownTestEnvironment());
-
-describe("promise-any test", () => {
-    const durableTestRunner = new LocalDurableTestRunner({
-        handlerFunction: handler,
-        skipTime: true,
-    });
-
-    it("should complete all promises", async () => {
-        const execution = await durableTestRunner.run();
-
-        expect(execution.getCompletedOperations()).toHaveLength(4)
-        console.log(execution.getResult());
-    });
-
-    it("should return expected result", async () => {
-        const execution = await durableTestRunner.run();
-
-        const result = execution.getResult() as PromiseSettledResult<any>[];
-
-        expect(result.filter(e => e.status! === "fulfilled")).toHaveLength(2);
-        expect(result.filter(e => e.status! === "rejected")).toHaveLength(1);
-    });
-
-    it("should fail if all promises fail - failure case", async () => {
-        durableTestRunner.getOperationByIndex(1).mockRejectedValue(new Error("ERROR"));
-        durableTestRunner.getOperationByIndex(2).mockRejectedValue(new Error("ERROR"));
-
-        const execution = await durableTestRunner.run();
-
-        expect(execution.getError()).toBeDefined();
-    })
-});
