+++ packages/aws-durable-execution-sdk-js-examples/src/examples/promise-any.ts
@@ -1,5 +1,6 @@
 import {
   DurableContext,
+  StepConfig,
   withDurableExecution,
 } from "@aws/durable-execution-sdk-js";
 import { ExampleConfig } from "../types";
@@ -9,13 +10,32 @@ export const config: ExampleConfig = {
   description: "Waiting for the first successful promise",
 };
 
+const stepConfig: StepConfig<any> = {
+  retryStrategy: (_error, attemptCount) => {
+    return {
+      shouldRetry: attemptCount < 3,
+      delaySeconds: 1,
+    };
+  },
+};
+
 export const handler = withDurableExecution(
   async (event: any, context: DurableContext) => {
     const promise1 = context.step(async () => {
       throw new Error("failure 1");
-    });
-    const promise2 = context.step(async () => "first success");
-    const promise3 = context.step(async () => "second success");
+    }, stepConfig);
+    const promise2 = context.step(async () => {
+      if (event.shouldFail) {
+        throw new Error("ERROR 1");
+      }
+      return "first success";
+    }, stepConfig);
+    const promise3 = context.step(async () => {
+      if (event.shouldFail) {
+        throw new Error("ERROR 2");
+      }
+      return "second success";
+    }, stepConfig);
 
     const result = await context.promise.any([promise1, promise2, promise3]);
