+++ packages/aws-durable-execution-sdk-js-testing/src/checkpoint-server/storage/__tests__/checkpoint-manager.test.ts
@@ -202,332 +202,11 @@ describe("CheckpointManager", () => {
       storage.completeOperation({
         Id: initialOperation.operation.Id,
         Action: OperationAction.SUCCEED,
+        Type: OperationType.EXECUTION,
       });
 
       expect(storage.hasOperation(initialOperation.operation.Id)).toBe(true);
     });
-  });
-
-  describe("completeOperation", () => {
-    it("should update operation status and add sequence", () => {
-      // Initialize with an operation
-      const initialOperation = storage.initialize();
-
-      // Complete the operation
-      const { operation } = storage.completeOperation({
-        Id: initialOperation.operation.Id,
-        Action: OperationAction.SUCCEED,
-      });
-
-      expect(operation).toBeDefined();
-      expect(operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(operation.EndTimestamp).toBeInstanceOf(Date);
-    });
-
-    it("should update operation with new step details", () => {
-      // Initialize storage
-      storage.initialize();
-
-      // Register a step operation
-      storage.registerUpdate(
-        {
-          Id: "new-id",
-          Action: OperationAction.START,
-          Type: OperationType.STEP,
-        },
-        mockInvocationId,
-      );
-
-      // Complete the operation
-      const { operation } = storage.completeOperation({
-        Id: "new-id",
-        Action: OperationAction.SUCCEED,
-        Payload: "new payload",
-      });
-
-      expect(operation).toBeDefined();
-      expect(operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(operation.EndTimestamp).toBeInstanceOf(Date);
-      expect(operation.StepDetails?.Result).toEqual("new payload");
-    });
-
-    it("should not modify original operation object", () => {
-      // Initialize with an operation
-      const initialOperation = storage.initialize();
-
-      // Complete the operation
-      const { operation } = storage.completeOperation({
-        Id: initialOperation.operation.Id,
-        Action: OperationAction.SUCCEED,
-      });
-
-      // Check that the original operation object is not modified
-      expect(initialOperation).not.toEqual(operation);
-    });
-
-    it("should throw error undefined for non-existent operation id", () => {
-      expect(() =>
-        storage.completeOperation({
-          Id: "non-existent-id",
-        }),
-      ).toThrow("Could not find operation");
-    });
-
-    it("should update operation with new context details", () => {
-      // Initialize storage
-      storage.initialize();
-
-      // Register a step operation
-      storage.registerUpdate(
-        {
-          Id: "new-id",
-          Action: OperationAction.START,
-          Type: OperationType.CONTEXT,
-          ContextOptions: {
-            ReplayChildren: true,
-          },
-        },
-        mockInvocationId,
-      );
-
-      // Complete the operation
-      const { operation } = storage.completeOperation({
-        Id: "new-id",
-        Action: OperationAction.SUCCEED,
-        Payload: "new payload",
-      });
-
-      expect(operation).toBeDefined();
-      expect(operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(operation.EndTimestamp).toBeInstanceOf(Date);
-      expect(operation.ContextDetails?.ReplayChildren).toBe(true);
-      expect(operation.ContextDetails?.Result).toEqual("new payload");
-    });
-  });
-
-  describe("registerUpdate", () => {
-    it("should throw error if id is missing", () => {
-      const updateWithoutId: OperationUpdate = {
-        Type: OperationType.STEP,
-        Action: OperationAction.START,
-      };
-
-      expect(() =>
-        storage.registerUpdate(updateWithoutId, mockInvocationId),
-      ).toThrow("Missing Id in update");
-    });
-
-    it("should create and register a new STEP operation", () => {
-      const update: OperationUpdate = {
-        Id: "step-id",
-        Action: OperationAction.START,
-        Name: "test-step",
-        Type: OperationType.STEP,
-      };
-
-      const result = storage.registerUpdate(update, mockInvocationId);
-
-      expect(result).toBeDefined();
-      expect(result.operation.Id).toBe("step-id");
-      expect(result.operation.Name).toBe("test-step");
-      expect(result.operation.Type).toBe(OperationType.STEP);
-      expect(result.operation.Status).toBe(OperationStatus.STARTED);
-      expect(result.operation.StartTimestamp).toBeInstanceOf(Date);
-      expect(result.operation.StepDetails).toBeUndefined();
-
-      expect(storage.operationDataMap.get("step-id")).toBe(result);
-    });
-
-    it("should create and register a new CONTEXT operation", () => {
-      const update: OperationUpdate = {
-        Id: "CONTEXT-id",
-        Name: "test-step",
-        Action: OperationAction.START,
-        Type: OperationType.CONTEXT,
-        ContextOptions: {
-          ReplayChildren: true,
-        },
-      };
-
-      const result = storage.registerUpdate(update, mockInvocationId);
-
-      expect(result).toBeDefined();
-      expect(result.operation.Id).toBe("CONTEXT-id");
-      expect(result.operation.Name).toBe("test-step");
-      expect(result.operation.Type).toBe(OperationType.CONTEXT);
-      expect(result.operation.Status).toBe(OperationStatus.STARTED);
-      expect(result.operation.StartTimestamp).toBeInstanceOf(Date);
-      expect(result.operation.StepDetails).toBeUndefined();
-      expect(result.operation.ContextDetails?.ReplayChildren).toBe(true);
-
-      expect(storage.operationDataMap.get("CONTEXT-id")).toBe(result);
-    });
-
-    it("should update operation with new step details", () => {
-      // Initialize storage
-      storage.initialize();
-
-      // Register a step operation
-      storage.registerUpdate(
-        {
-          Id: "new-id",
-          Action: OperationAction.START,
-          Type: OperationType.STEP,
-        },
-        mockInvocationId,
-      );
-
-      // Register the operation
-      const { operation } = storage.registerUpdate(
-        {
-          Id: "new-id",
-          Action: OperationAction.SUCCEED,
-          Payload: "new payload",
-        },
-        mockInvocationId,
-      );
-
-      expect(operation).toBeDefined();
-      expect(operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(operation.EndTimestamp).toBeInstanceOf(Date);
-      expect(operation.StepDetails?.Result).toEqual("new payload");
-    });
-
-    it("should create STEP operation with payload but not store result in registerUpdate", () => {
-      const update: OperationUpdate = {
-        Id: "step-with-payload",
-        Name: "process-data",
-        Type: OperationType.STEP,
-        Action: OperationAction.START,
-        Payload: JSON.stringify({ processed: true, value: 42 }),
-      };
-
-      const result = storage.registerUpdate(update, mockInvocationId);
-
-      // Result should not be populated in registerUpdate, only in completeOperation
-      expect(result.operation.StepDetails).toBeUndefined();
-      expect(result.operation.Status).toBe(OperationStatus.STARTED);
-    });
-
-    it("should handle STEP operation with undefined payload", () => {
-      const update: OperationUpdate = {
-        Id: "step-no-payload",
-        Type: OperationType.STEP,
-        Action: OperationAction.START,
-        Payload: undefined,
-      };
-
-      const result = storage.registerUpdate(update, mockInvocationId);
-
-      // Result should not be populated in registerUpdate, only in completeOperation
-      expect(result.operation.StepDetails).toBeUndefined();
-    });
-
-    it("should create and register a new WAIT operation with scheduled timestamp", () => {
-      const update: OperationUpdate = {
-        Id: "wait-id",
-        Name: "test-wait",
-        Type: OperationType.WAIT,
-        Action: OperationAction.START,
-        WaitOptions: {
-          WaitSeconds: 5,
-        },
-      };
-
-      const result = storage.registerUpdate(update, mockInvocationId);
-
-      expect(result).toBeDefined();
-      expect(result.operation.Id).toBe("wait-id");
-      expect(result.operation.Type).toBe(OperationType.WAIT);
-      expect(
-        result.operation.WaitDetails?.ScheduledEndTimestamp,
-      ).toBeInstanceOf(Date);
-
-      // Make sure waitDetails exists and has a scheduledEndTimestamp
-      expect(result.operation.WaitDetails).toBeDefined();
-      expect(result.operation.WaitDetails?.ScheduledEndTimestamp).toBeDefined();
-
-      // Since we've already verified the timestamp exists and is a Date (earlier in this test),
-      // we can now verify it's approximately 5 seconds in the future
-
-      // Create a reference time just before we call the function to compare against
-      const now = new Date().getTime();
-
-      const timestamp = result.operation.WaitDetails?.ScheduledEndTimestamp;
-      expect(timestamp).toBeInstanceOf(Date);
-
-      const scheduledTime = timestamp instanceof Date ? timestamp.getTime() : 0;
-
-      // Allow 1 second tolerance for test execution time
-      expect(scheduledTime).toBeGreaterThan(now + 4000);
-      expect(scheduledTime).toBeLessThan(now + 6000);
-    });
-
-    it("should return existing operation if it has waitDetails with scheduledEndTimestamp", () => {
-      // First register a wait operation
-      const waitUpdate: OperationUpdate = {
-        Id: "wait-id",
-        Type: OperationType.WAIT,
-        Action: OperationAction.START,
-        WaitOptions: {
-          WaitSeconds: 5,
-        },
-      };
-      const firstResult = storage.registerUpdate(waitUpdate, mockInvocationId);
-
-      // Try to register it again
-      const secondResult = storage.registerUpdate(waitUpdate, mockInvocationId);
-
-      expect(secondResult).toBe(firstResult);
-    });
-
-    it("should add operation to pending updates", async () => {
-      // Setup to capture the pending updates
-      const pendingUpdatesPromise = storage.getPendingCheckpointUpdates();
-
-      // Register an update
-      const update: OperationUpdate = {
-        Id: "test-update",
-        Type: OperationType.STEP,
-        Action: OperationAction.START,
-      };
-      storage.registerUpdate(update, mockInvocationId);
-
-      // Get the pending updates
-      const pendingUpdates = await pendingUpdatesPromise;
-
-      expect(pendingUpdates).toHaveLength(1);
-      expect(pendingUpdates[0].update?.Id).toBe("test-update");
-    });
-
-    it.each([OperationAction.SUCCEED, OperationAction.FAIL])(
-      "should add operation as completed if the action is %s",
-      (action) => {
-        const update: OperationUpdate = {
-          Id: "step-id",
-          Name: "test-step",
-          Type: OperationType.STEP,
-          Action: action,
-        };
-
-        const result = storage.registerUpdate(update, mockInvocationId);
-
-        expect(result).toBeDefined();
-        expect(result.operation.Id).toBe("step-id");
-        expect(result.operation.Name).toBe("test-step");
-        expect(result.operation.Type).toBe(OperationType.STEP);
-        expect(result.operation.Status).toBe(
-          action === OperationAction.SUCCEED
-            ? OperationStatus.SUCCEEDED
-            : OperationStatus.FAILED,
-        );
-        expect(result.operation.StartTimestamp).toBeInstanceOf(Date);
-        expect(result.operation.StepDetails).toEqual({});
-        expect(result.operation.EndTimestamp).toBeInstanceOf(Date);
-
-        expect(storage.operationDataMap.get("step-id")).toBe(result);
-      },
-    );
   });
 
   describe("retry functionality", () => {
@@ -548,6 +227,7 @@ describe("CheckpointManager", () => {
           Id: "retry-step",
           Action: OperationAction.RETRY,
           Payload: "retry result",
+          Type: OperationType.STEP,
         });
 
         expect(result.operation.StepDetails?.Attempt).toBe(2); // 1 from register + 1 from complete
@@ -570,6 +250,7 @@ describe("CheckpointManager", () => {
           Id: "success-step",
           Action: OperationAction.SUCCEED,
           Payload: "final result",
+          Type: OperationType.STEP,
         });
 
         expect(result.operation.StepDetails?.Attempt).toBe(1); // Preserved from register
@@ -592,6 +273,7 @@ describe("CheckpointManager", () => {
         const result = storage.completeOperation({
           Id: "new-retry-step",
           Action: OperationAction.RETRY,
+          Type: OperationType.STEP,
         });
 
         expect(result.operation.StepDetails?.Attempt).toBe(1); // 0 + 1
@@ -690,6 +372,7 @@ describe("CheckpointManager", () => {
         const result = storage.completeOperation({
           Id: "step-with-error",
           Action: OperationAction.FAIL,
+          Type: OperationType.STEP,
           Error: {
             ErrorMessage: "Step execution failed",
             ErrorType: "StepExecutionError",
@@ -723,6 +406,7 @@ describe("CheckpointManager", () => {
           Id: "step-partial",
           Action: OperationAction.FAIL,
           Payload: '{"partial": "data"}',
+          Type: OperationType.STEP,
           Error: {
             ErrorMessage: "Partial failure occurred",
             ErrorType: "PartialFailure",
@@ -756,6 +440,7 @@ describe("CheckpointManager", () => {
         const result = storage.completeOperation({
           Id: "retry-step-error",
           Action: OperationAction.RETRY,
+          Type: OperationType.STEP,
           Error: {
             ErrorMessage: "Retry attempt failed",
             ErrorType: "RetryError",
@@ -794,6 +479,7 @@ describe("CheckpointManager", () => {
         const result = storage.completeOperation({
           Id: "manual-step",
           Action: OperationAction.FAIL,
+          Type: OperationType.STEP,
           Error: {
             ErrorMessage: "New error",
             ErrorType: "NewError",
@@ -963,6 +649,7 @@ describe("CheckpointManager", () => {
 
         // Complete the operation with new result
         const result = storage.completeOperation({
+          Type: OperationType.STEP,
           Id: "context-to-complete",
           Action: OperationAction.SUCCEED,
           Payload: '{"final": "result"}',
@@ -993,6 +680,7 @@ describe("CheckpointManager", () => {
         const result = storage.completeOperation({
           Id: "context-to-fail",
           Action: OperationAction.FAIL,
+          Type: OperationType.STEP,
           Error: {
             ErrorMessage: "Context execution failed",
             ErrorType: "ContextExecutionError",
@@ -1025,6 +713,7 @@ describe("CheckpointManager", () => {
           Id: "context-partial",
           Action: OperationAction.FAIL,
           Payload: '{"partial": "data"}',
+          Type: OperationType.STEP,
           Error: {
             ErrorMessage: "Partial failure occurred",
             ErrorType: "PartialFailure",
@@ -1066,6 +755,7 @@ describe("CheckpointManager", () => {
         const result = storage.completeOperation({
           Id: "manual-context",
           Action: OperationAction.SUCCEED,
+          Type: OperationType.STEP,
           Payload: '{"new": "result"}',
         });
 
@@ -1175,6 +865,7 @@ describe("CheckpointManager", () => {
           Id: "test-callback-op",
           Type: OperationType.CALLBACK,
           Status: OperationStatus.STARTED,
+          StartTimestamp: new Date(),
           CallbackDetails: {
             CallbackId: mockCallbackId,
           },
@@ -1268,6 +959,7 @@ describe("CheckpointManager", () => {
           Id: "test-op",
           Type: OperationType.CALLBACK,
           Status: OperationStatus.STARTED,
+          StartTimestamp: new Date(),
         },
         update: {
           Id: "test-op",
@@ -1449,852 +1141,6 @@ describe("CheckpointManager", () => {
       expect(results[0].operation.Id).toBe("op-z");
       expect(results[1].operation.Id).toBe("op-a");
       expect(results[2].operation.Id).toBe("op-m");
-    });
-  });
-
-  describe("updateOperation", () => {
-    it("should update operation for WAIT and update history", () => {
-      const initialOperation = storage.initialize();
-
-      const newOperationData = {
-        Type: OperationType.WAIT,
-        Status: OperationStatus.SUCCEEDED,
-        EndTimestamp: new Date(),
-      };
-
-      const result = storage.updateOperation(
-        initialOperation.operation.Id,
-        newOperationData,
-      );
-
-      // Should return the updated CheckpointOperation
-      expect(result.operation.Id).toBe(initialOperation.operation.Id);
-      expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED); // Updated status
-      expect(result.operation.EndTimestamp).toEqual(
-        newOperationData.EndTimestamp,
-      );
-
-      expect(result.events).toEqual([
-        {
-          EventType: "ExecutionStarted",
-          SubType: undefined,
-          EventId: 1,
-          Id: "mocked-uuid",
-          Name: undefined,
-          EventTimestamp: expect.any(Date),
-          ParentId: undefined,
-          ExecutionStartedDetails: {
-            Input: {
-              Payload: "{}",
-            },
-            ExecutionTimeout: undefined,
-          },
-        },
-        {
-          EventType: "WaitSucceeded",
-          SubType: undefined,
-          EventId: 2,
-          Id: "mocked-uuid",
-          Name: undefined,
-          EventTimestamp: expect.any(Date),
-          ParentId: undefined,
-          WaitSucceededDetails: { Duration: undefined },
-        },
-      ]);
-    });
-
-    describe("CHAINED_INVOKE operation updates", () => {
-      describe("valid operation updates", () => {
-        it("should update CHAINED_INVOKE operation with SUCCEEDED status and create history event", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.SUCCEEDED,
-            EndTimestamp: new Date(),
-            ChainedInvokeDetails: {
-              Result: '{"success": true, "data": "test-result"}',
-            },
-          };
-
-          const result = storage.updateOperation("invoke-op", newOperationData);
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe("invoke-op");
-          expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-          expect(result.operation.ChainedInvokeDetails?.Result).toBe(
-            '{"success": true, "data": "test-result"}',
-          );
-
-          // Should have added InvokeSucceeded event
-          expect(result.events).toHaveLength(2); // Start event + Success event
-          const successEvent = result.events[1];
-          expect(successEvent.EventType).toBe("ChainedInvokeSucceeded");
-          expect(successEvent.ChainedInvokeSucceededDetails).toEqual({
-            Result: {
-              Payload: '{"success": true, "data": "test-result"}',
-            },
-            Error: undefined,
-          });
-        });
-
-        it("should update CHAINED_INVOKE operation with FAILED status", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-fail-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke-fail",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.FAILED,
-            EndTimestamp: new Date(),
-            ChainedInvokeDetails: {
-              Error: {
-                ErrorType: "InvokeError",
-                ErrorMessage: "Function failed",
-              },
-            },
-          };
-
-          const result = storage.updateOperation(
-            "invoke-fail-op",
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe("invoke-fail-op");
-          expect(result.operation.Status).toBe(OperationStatus.FAILED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-          expect(result.operation.ChainedInvokeDetails?.Error).toEqual({
-            ErrorType: "InvokeError",
-            ErrorMessage: "Function failed",
-          });
-
-          // Should have added ChainedInvokeFailed event
-          expect(result.events).toHaveLength(2); // Start event + Failed event
-          const failedEvent = result.events[1];
-          expect(failedEvent.EventType).toBe("ChainedInvokeFailed");
-          expect(failedEvent.ChainedInvokeFailedDetails).toEqual({
-            Result: undefined,
-            Error: {
-              Payload: {
-                ErrorType: "InvokeError",
-                ErrorMessage: "Function failed",
-              },
-            },
-          });
-        });
-
-        it("should update CHAINED_INVOKE operation with CANCELLED status", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-cancel-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke-cancel",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.CANCELLED,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            "invoke-cancel-op",
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe("invoke-cancel-op");
-          expect(result.operation.Status).toBe(OperationStatus.CANCELLED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-
-          // Should have added InvokeCancelled event
-          expect(result.events).toHaveLength(2); // Start event + Cancelled event
-          const cancelledEvent = result.events[1];
-          expect(cancelledEvent.EventType).toBe("ChainedInvokeCancelled");
-          expect(cancelledEvent.ChainedInvokeStoppedDetails).toEqual({
-            Result: undefined,
-            Error: undefined,
-          });
-        });
-
-        it("should update CHAINED_INVOKE operation with TIMED_OUT status", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-timeout-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke-timeout",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.TIMED_OUT,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            "invoke-timeout-op",
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe("invoke-timeout-op");
-          expect(result.operation.Status).toBe(OperationStatus.TIMED_OUT);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-
-          // Should have added InvokeTimedOut event
-          expect(result.events).toHaveLength(2); // Start event + TimedOut event
-          const timedOutEvent = result.events[1];
-          expect(timedOutEvent.EventType).toBe("ChainedInvokeTimedOut");
-          expect(timedOutEvent.ChainedInvokeTimedOutDetails).toEqual({
-            Result: undefined,
-            Error: undefined,
-          });
-        });
-
-        it("should update CHAINED_INVOKE operation without ChainedInvokeDetails", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-no-details-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke-no-details",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.SUCCEEDED,
-            EndTimestamp: new Date(),
-            // No ChainedInvokeDetails provided
-          };
-
-          const result = storage.updateOperation(
-            "invoke-no-details-op",
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe("invoke-no-details-op");
-          expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED);
-
-          // Should have added InvokeSucceeded event with undefined Result and Error
-          expect(result.events).toHaveLength(2); // Start event + Success event
-          const successEvent = result.events[1];
-          expect(successEvent.EventType).toBe("ChainedInvokeSucceeded");
-          expect(successEvent.ChainedInvokeSucceededDetails).toEqual({
-            Result: undefined,
-            Error: undefined,
-          });
-        });
-
-        it("should handle CHAINED_INVOKE operation with empty Result and Error strings", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-empty-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke-empty",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.SUCCEEDED,
-            EndTimestamp: new Date(),
-            ChainedInvokeDetails: {
-              Result: "", // Empty string
-              Error: {
-                ErrorType: "",
-                ErrorMessage: "",
-              },
-            },
-          };
-
-          const result = storage.updateOperation(
-            "invoke-empty-op",
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe("invoke-empty-op");
-          expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED);
-          expect(result.operation.ChainedInvokeDetails?.Result).toBe("");
-          expect(result.operation.ChainedInvokeDetails?.Error).toEqual({
-            ErrorMessage: "",
-            ErrorType: "",
-          });
-
-          // Should have added InvokeSucceeded event with empty string payloads
-          expect(result.events).toHaveLength(2); // Start event + Success event
-          const successEvent = result.events[1];
-          expect(successEvent.EventType).toBe("ChainedInvokeSucceeded");
-          expect(successEvent.ChainedInvokeSucceededDetails).toEqual({
-            Result: {
-              Payload: "",
-            },
-            Error: {
-              Payload: {
-                ErrorMessage: "",
-                ErrorType: "",
-              },
-            },
-          });
-        });
-      });
-
-      describe("invalid operation updates", () => {
-        it("should not update CHAINED_INVOKE operation with both Result and Error", () => {
-          storage.initialize();
-
-          // Register an CHAINED_INVOKE operation first
-          const invokeUpdate: OperationUpdate = {
-            Id: "invoke-both-op",
-            Type: OperationType.CHAINED_INVOKE,
-            Action: OperationAction.START,
-            Name: "test-invoke-both",
-          };
-          storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-          const newOperationData = {
-            Type: OperationType.CHAINED_INVOKE,
-            Status: OperationStatus.FAILED,
-            EndTimestamp: new Date(),
-            ChainedInvokeDetails: {
-              Result: '{"partial": "data"}',
-              Error: {
-                ErrorType: "PartialError",
-                ErrorMessage: "Partial failure",
-              },
-            },
-          };
-
-          expect(() =>
-            storage.updateOperation("invoke-both-op", newOperationData),
-          ).toThrow(
-            "Could not update operation with both Result and Error in details.",
-          );
-        });
-
-        it.each([
-          OperationStatus.PENDING,
-          OperationStatus.READY,
-          OperationStatus.STOPPED,
-          OperationStatus.STARTED,
-        ])(
-          `should throw error for invalid CHAINED_INVOKE status %s`,
-          (status) => {
-            storage.initialize();
-
-            // Register an CHAINED_INVOKE operation first
-            const invokeUpdate: OperationUpdate = {
-              Id: "invoke-invalid-op",
-              Type: OperationType.CHAINED_INVOKE,
-              Action: OperationAction.START,
-              Name: "test-invoke-invalid",
-            };
-            storage.registerUpdate(invokeUpdate, mockInvocationId);
-
-            const newOperationData = {
-              Type: OperationType.CHAINED_INVOKE,
-              Status: status,
-              EndTimestamp: new Date(),
-            };
-
-            expect(() => {
-              storage.updateOperation("invoke-invalid-op", newOperationData);
-            }).toThrow(`Invalid status update for CHAINED_INVOKE: ${status}`);
-          },
-        );
-      });
-    });
-
-    describe("EXECUTION operation updates", () => {
-      describe("valid operation updates", () => {
-        it("should update EXECUTION operation with SUCCEEDED status and create history event", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.SUCCEEDED,
-            EndTimestamp: new Date(),
-            ExecutionDetails: {
-              InputPayload: '{"success": true, "data": "test-result"}',
-            },
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-            '{"success": true, "data": "test-result"}',
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-          expect(result.operation.ExecutionDetails?.InputPayload).toBe(
-            '{"success": true, "data": "test-result"}',
-          );
-
-          // Should have added ExecutionSucceeded event
-          expect(result.events).toHaveLength(2); // Start event + Success event
-          const successEvent = result.events[1];
-          expect(successEvent.EventType).toBe("ExecutionSucceeded");
-          expect(successEvent.ExecutionSucceededDetails).toEqual({
-            Result: {
-              Payload: '{"success": true, "data": "test-result"}',
-            },
-            Error: undefined,
-          });
-        });
-
-        it("should update EXECUTION operation with FAILED status", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.FAILED,
-            EndTimestamp: new Date(),
-            ExecutionDetails: {
-              InputPayload: "{}",
-            },
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-            undefined,
-            {
-              ErrorType: "ExecutionError",
-              ErrorMessage: "Execution failed",
-            },
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.FAILED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-
-          // Should have added ExecutionFailed event
-          expect(result.events).toHaveLength(2); // Start event + Failed event
-          const failedEvent = result.events[1];
-          expect(failedEvent.EventType).toBe("ExecutionFailed");
-          expect(failedEvent.ExecutionFailedDetails).toEqual({
-            Result: undefined,
-            Error: {
-              Payload: {
-                ErrorType: "ExecutionError",
-                ErrorMessage: "Execution failed",
-              },
-            },
-          });
-        });
-
-        it("should update EXECUTION operation with STOPPED status", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.STOPPED,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.STOPPED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-
-          // Should have added ExecutionStopped event
-          expect(result.events).toHaveLength(2); // Start event + Stopped event
-          const stoppedEvent = result.events[1];
-          expect(stoppedEvent.EventType).toBe("ExecutionStopped");
-          expect(stoppedEvent.ExecutionStoppedDetails).toEqual({
-            Result: undefined,
-            Error: undefined,
-          });
-        });
-
-        it("should update EXECUTION operation with TIMED_OUT status", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.TIMED_OUT,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.TIMED_OUT);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-
-          // Should have added ExecutionTimedOut event
-          expect(result.events).toHaveLength(2); // Start event + TimedOut event
-          const timedOutEvent = result.events[1];
-          expect(timedOutEvent.EventType).toBe("ExecutionTimedOut");
-          expect(timedOutEvent.ExecutionTimedOutDetails).toEqual({
-            Result: undefined,
-            Error: undefined,
-          });
-        });
-
-        it("should update EXECUTION operation with STARTED status", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.STARTED,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.STARTED);
-          expect(result.operation.EndTimestamp).toEqual(
-            newOperationData.EndTimestamp,
-          );
-
-          // Should have added ExecutionStarted event
-          expect(result.events).toHaveLength(2); // Start event + Started event
-          const startedEvent = result.events[1];
-          expect(startedEvent.EventType).toBe("ExecutionStarted");
-          expect(startedEvent.ExecutionStartedDetails).toEqual({
-            Result: undefined,
-            Error: undefined,
-          });
-        });
-
-        it("should update EXECUTION operation with both result and error", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.FAILED,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-            '{"partial": "data"}',
-            {
-              ErrorType: "PartialError",
-              ErrorMessage: "Partial failure",
-            },
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.FAILED);
-
-          // Should have added ExecutionFailed event with both Result and Error
-          expect(result.events).toHaveLength(2); // Start event + Failed event
-          const failedEvent = result.events[1];
-          expect(failedEvent.EventType).toBe("ExecutionFailed");
-          expect(failedEvent.ExecutionFailedDetails).toEqual({
-            Result: {
-              Payload: '{"partial": "data"}',
-            },
-            Error: {
-              Payload: {
-                ErrorType: "PartialError",
-                ErrorMessage: "Partial failure",
-              },
-            },
-          });
-        });
-
-        it("should handle EXECUTION operation with empty Result and Error strings", () => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: OperationStatus.SUCCEEDED,
-            EndTimestamp: new Date(),
-          };
-
-          const result = storage.updateOperation(
-            initialOperation.operation.Id,
-            newOperationData,
-            "", // Empty string
-            {
-              ErrorType: "",
-              ErrorMessage: "",
-            },
-          );
-
-          // Should return the updated CheckpointOperation
-          expect(result.operation.Id).toBe(initialOperation.operation.Id);
-          expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED);
-
-          // Should have added ExecutionSucceeded event with empty string payloads
-          expect(result.events).toHaveLength(2); // Start event + Success event
-          const successEvent = result.events[1];
-          expect(successEvent.EventType).toBe("ExecutionSucceeded");
-          expect(successEvent.ExecutionSucceededDetails).toEqual({
-            Result: {
-              Payload: "",
-            },
-            Error: {
-              Payload: {
-                ErrorMessage: "",
-                ErrorType: "",
-              },
-            },
-          });
-        });
-      });
-
-      describe("invalid operation updates", () => {
-        it.each([
-          OperationStatus.PENDING,
-          OperationStatus.READY,
-          OperationStatus.CANCELLED,
-        ])(`should throw error for invalid EXECUTION status %s`, (status) => {
-          const initialOperation = storage.initialize();
-
-          const newOperationData = {
-            Type: OperationType.EXECUTION,
-            Status: status,
-            EndTimestamp: new Date(),
-          };
-
-          expect(() => {
-            storage.updateOperation(
-              initialOperation.operation.Id,
-              newOperationData,
-            );
-          }).toThrow(`Invalid status update for EXECUTION: ${status}`);
-        });
-      });
-    });
-
-    it("should update existing operation and return the updated CheckpointOperation", () => {
-      const initialOperation = storage.initialize();
-
-      const newOperationData = {
-        Status: OperationStatus.SUCCEEDED,
-        EndTimestamp: new Date(),
-      };
-
-      const result = storage.updateOperation(
-        initialOperation.operation.Id,
-        newOperationData,
-      );
-
-      // Should return the updated CheckpointOperation
-      expect(result.operation.Id).toBe(initialOperation.operation.Id);
-      expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED); // Updated status
-      expect(result.operation.EndTimestamp).toEqual(
-        newOperationData.EndTimestamp,
-      );
-
-      // And the stored operation should be the same as returned
-      const storedOperation = storage.operationDataMap.get(
-        initialOperation.operation.Id,
-      );
-      expect(storedOperation).toBe(result);
-      expect(storedOperation?.operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(storedOperation?.operation.EndTimestamp).toEqual(
-        newOperationData.EndTimestamp,
-      );
-    });
-
-    it("should preserve existing operation properties when updating", () => {
-      const initialOperation = storage.initialize();
-      const originalType = initialOperation.operation.Type;
-      const originalExecutionDetails =
-        initialOperation.operation.ExecutionDetails;
-
-      const newOperationData = {
-        Status: OperationStatus.FAILED,
-        SomeNewField: "new-value",
-      };
-
-      storage.updateOperation(initialOperation.operation.Id, newOperationData);
-
-      const storedOperation = storage.operationDataMap.get(
-        initialOperation.operation.Id,
-      );
-      expect(storedOperation?.operation.Type).toBe(originalType);
-      expect(storedOperation?.operation.ExecutionDetails).toBe(
-        originalExecutionDetails,
-      );
-      expect(storedOperation?.operation.Status).toBe(OperationStatus.FAILED);
-      expect(
-        (storedOperation?.operation as Record<string, unknown>).SomeNewField,
-      ).toBe("new-value");
-    });
-
-    it("should handle partial operation updates", () => {
-      storage.initialize();
-
-      // Register a step operation first
-      const stepUpdate: OperationUpdate = {
-        Id: "step-to-update",
-        Type: OperationType.STEP,
-        Action: OperationAction.START,
-        Name: "test-step",
-      };
-      storage.registerUpdate(stepUpdate, mockInvocationId);
-
-      const partialUpdate = {
-        Status: OperationStatus.SUCCEEDED,
-      };
-
-      const result = storage.updateOperation("step-to-update", partialUpdate);
-
-      expect(result.operation.Id).toBe("step-to-update");
-      expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED); // Updated
-
-      // Check stored operation is the same as returned
-      const storedOperation = storage.operationDataMap.get("step-to-update");
-      expect(storedOperation).toBe(result);
-      expect(storedOperation?.operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(storedOperation?.operation.Name).toBe("test-step"); // Preserved
-      expect(storedOperation?.operation.Type).toBe(OperationType.STEP); // Preserved
-    });
-
-    it("should throw error for non-existent operation", () => {
-      storage.initialize();
-
-      expect(() => {
-        storage.updateOperation("non-existent-id", {
-          Status: OperationStatus.SUCCEEDED,
-        });
-      }).toThrow("Could not find operation");
-    });
-
-    it("should handle complex operation updates", () => {
-      storage.initialize();
-
-      // Register a complex operation
-      const complexUpdate: OperationUpdate = {
-        Id: "complex-op",
-        Action: OperationAction.START,
-        Type: OperationType.STEP,
-        Name: "complex-step",
-        Payload: "initial-payload",
-      };
-      storage.registerUpdate(complexUpdate, mockInvocationId);
-
-      const updateData = {
-        Status: OperationStatus.SUCCEEDED,
-        EndTimestamp: new Date(),
-        StepDetails: {
-          Result: "final-result",
-          Attempt: 1,
-        },
-      };
-
-      const result = storage.updateOperation("complex-op", updateData);
-
-      // Updated operation returned
-      expect(result.operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(result.operation.StepDetails).toEqual({
-        Result: "final-result",
-        Attempt: 1,
-      });
-      expect(result.operation.EndTimestamp).toEqual(updateData.EndTimestamp);
-
-      // Stored operation should be the same as returned
-      const storedOperation = storage.operationDataMap.get("complex-op");
-      expect(storedOperation).toBe(result);
-      expect(storedOperation?.operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(storedOperation?.operation.StepDetails).toEqual({
-        Result: "final-result",
-        Attempt: 1,
-      });
-      expect(storedOperation?.operation.EndTimestamp).toEqual(
-        updateData.EndTimestamp,
-      );
-    });
-
-    it("should handle updates with undefined or null values", () => {
-      const initialOperation = storage.initialize();
-
-      const updateWithUndefined = {
-        Status: OperationStatus.SUCCEEDED,
-        SomeField: undefined,
-        AnotherField: null,
-      };
-
-      storage.updateOperation(
-        initialOperation.operation.Id,
-        updateWithUndefined,
-      );
-
-      const storedOperation = storage.operationDataMap.get(
-        initialOperation.operation.Id,
-      );
-      expect(storedOperation?.operation.Status).toBe(OperationStatus.SUCCEEDED);
-      expect(
-        (storedOperation?.operation as Record<string, unknown>).SomeField,
-      ).toBeUndefined();
-      expect(
-        (storedOperation?.operation as Record<string, unknown>).AnotherField,
-      ).toBeNull();
     });
   });
